+++ cryptofeed/backends/arctic.py
@@ -8,8 +8,9 @@ import arctic
 import pandas as pd
 
 from cryptofeed.backends.backend import (BackendFundingCallback, BackendOpenInterestCallback,
-                                         BackendTickerCallback, BackendTradeCallback, BackendLiquidationsCallback)
-from cryptofeed.defines import FUNDING, OPEN_INTEREST, TICKER, TRADES, LIQUIDATIONS
+                                         BackendTickerCallback, BackendTradeCallback, BackendLiquidationsCallback,
+                                         BackendMarketInfoCallback, BackendTransactionsCallback)
+from cryptofeed.defines import FUNDING, OPEN_INTEREST, TICKER, TRADES, LIQUIDATIONS, MARKET_INFO, TRANSACTIONS
 
 
 class ArcticCallback:
@@ -65,3 +66,12 @@ class OpenInterestArctic(ArcticCallback, BackendOpenInterestCallback):
 
 class LiquidationsArctic(ArcticCallback, BackendLiquidationsCallback):
     default_key = LIQUIDATIONS
+
+
+class MarketInfoArctic(ArcticCallback, BackendMarketInfoCallback):
+    default_key = MARKET_INFO
+    
+
+class TransactionsArctic(ArcticCallback, BackendTransactionsCallback):
+    default_key = TRANSACTIONS
+