+++ cryptofeed/backends/elastic.py
@@ -14,7 +14,7 @@ from yapic import json
 from cryptofeed.backends._util import book_flatten
 from cryptofeed.backends.backend import (BackendBookCallback, BackendBookDeltaCallback, BackendFundingCallback,
                                          BackendOpenInterestCallback, BackendTickerCallback, BackendTradeCallback,
-                                         BackendLiquidationsCallback)
+                                         BackendLiquidationsCallback, BackendMarketInfoCallback, BackendTansactionsCallback)
 from cryptofeed.backends.http import HTTPCallback
 
 
@@ -100,3 +100,12 @@ class OpenInterestElastic(ElasticCallback, BackendOpenInterestCallback):
 
 class LiquidationsElastic(ElasticCallback, BackendLiquidationsCallback):
     default_index = 'liquidations'
+
+
+class MarketInfoElastic(ElasticCallback, BackendMarketInfoCallback):
+    default_index = 'market_info'
+
+
+class TransactionsElastic(ElasticCallback, BackendTransactionsCallback):
+    default_index = 'transactions'
+