+++ cryptofeed/feed.py
@@ -8,8 +8,8 @@ import uuid
 from collections import defaultdict
 
 from cryptofeed.callback import Callback
-from cryptofeed.defines import (ASK, BID, BOOK_DELTA, FUNDING, L2_BOOK, L3_BOOK,
-                                LIQUIDATIONS, OPEN_INTEREST, TICKER, TRADES, VOLUME, FUTURES_INDEX)
+from cryptofeed.defines import (ASK, BID, BOOK_DELTA, FUNDING, FUTURES_INDEX, L2_BOOK, L3_BOOK, LIQUIDATIONS,
+                                OPEN_INTEREST, MARKET_INFO, TICKER, TRADES, TRANSACTIONS, VOLUME)
 from cryptofeed.exceptions import BidAskOverlapping, UnsupportedDataFeed
 from cryptofeed.standards import feed_to_exchange, get_exchange_info, load_exchange_pair_mapping, pair_std_to_exchange
 from cryptofeed.util.book import book_delta, depth
@@ -18,7 +18,8 @@ from cryptofeed.util.book import book_delta, depth
 class Feed:
     id = 'NotImplemented'
 
-    def __init__(self, address, pairs=None, channels=None, config=None, callbacks=None, max_depth=None, book_interval=1000, snapshot_interval=False, checksum_validation=False, cross_check=False, origin=None):
+    def __init__(self, address, pairs=None, channels=None, config=None, callbacks=None, max_depth=None, book_interval=1000, snapshot_interval=False, checksum_validation=False, cross_check=False, origin=None,
+                 key_id=None):
         """
         max_depth: int
             Maximum number of levels per side to return in book updates
@@ -35,6 +36,8 @@ class Feed:
             checksums or provide message sequence numbers.
         origin: str
             Passed into websocket connect. Sets the origin header.
+        key_id: str
+            API key to query the feed, required when requesting supported coins/pairs.
         """
         self.hash = str(uuid.uuid4())
         self.uuid = f"{self.id}-{self.hash}"
@@ -51,7 +54,7 @@ class Feed:
         self.previous_book = defaultdict(dict)
         self.origin = origin
         self.checksum_validation = checksum_validation
-        load_exchange_pair_mapping(self.id)
+        load_exchange_pair_mapping(self.id, key_id)
 
         if config is not None and (pairs is not None or channels is not None):
             raise ValueError("Use config, or channels and pairs, not both")
@@ -68,15 +71,19 @@ class Feed:
 
         self.l3_book = {}
         self.l2_book = {}
-        self.callbacks = {TRADES: Callback(None),
-                          TICKER: Callback(None),
+        self.callbacks = {
+                          FUNDING: Callback(None),
+                          FUTURES_INDEX: Callback(None),
                           L2_BOOK: Callback(None),
                           L3_BOOK: Callback(None),
-                          VOLUME: Callback(None),
-                          FUNDING: Callback(None),
-                          OPEN_INTEREST: Callback(None),
                           LIQUIDATIONS: Callback(None),
-                          FUTURES_INDEX: Callback(None)}
+                          OPEN_INTEREST: Callback(None),
+                          MARKET_INFO: Callback(None),
+                          TICKER: Callback(None),
+                          TRADES: Callback(None),
+                          TRANSACTIONS: Callback(None),
+                          VOLUME: Callback(None)
+                          }
 
         if callbacks:
             for cb_type, cb_func in callbacks.items():
@@ -89,13 +96,16 @@ class Feed:
                 self.callbacks[key] = [callback]
 
     @classmethod
-    def info(cls) -> dict:
+    def info(cls, key_id: str = None) -> dict:
         """
         Return information about the Exchange - what trading pairs are supported, what data channels, etc
+
+        key_id: str
+            API key to query the feed, required when requesting supported coins/pairs.
         """
-        pairs, info = get_exchange_info(cls.id)
+        pairs, info = get_exchange_info(cls.id, key_id)
         data = {'pairs': list(pairs.keys()), 'channels': []}
-        for channel in (LIQUIDATIONS, OPEN_INTEREST, FUNDING, VOLUME, TICKER, L2_BOOK, L3_BOOK, TRADES, FUTURES_INDEX):
+        for channel in (FUNDING, FUTURES_INDEX, LIQUIDATIONS, L2_BOOK, L3_BOOK, OPEN_INTEREST, MARKET_INFO, TICKER, TRADES, TRANSACTIONS, VOLUME):
             try:
                 feed_to_exchange(cls.id, channel, silent=True)
                 data['channels'].append(channel)
@@ -103,7 +113,6 @@ class Feed:
                 pass
 
         data.update(info)
-
         return data
 
     async def book_callback(self, book: dict, book_type: str, pair: str, forced: bool, delta: dict, timestamp: float, receipt_timestamp: float):
@@ -194,5 +203,7 @@ class Feed:
 
 
 class RestFeed(Feed):
+    sleep_time = None
+
     async def message_handler(self):
         raise NotImplementedError