+++ cryptofeed/providers.py
@@ -1,0 +1,9 @@
+'''
+Copyright (C) 2017-2020  Bryant Moscon - bmoscon@gmail.com
+
+Please see the LICENSE file for the terms and conditions
+associated with this software.
+'''
+from cryptofeed.provider.coingecko import Coingecko
+from cryptofeed.provider.whale_alert import WhaleAlert
+